refactor: 优化前端的目录结构
