+++ src/panes/ChapterPane.vue
